refactor: playerStore 타입 분리 및 리팩토링
