+++ src/store/playerStore.ts
@@ -1,65 +1,10 @@
 import { setRepeatMode, setShuffle } from "@/apis/playbackApi";
 import { mapSdkTrackToLocalTrack } from "@/lib/spotifyMapper";
+import { PlayerState, RepeatMode } from "@/types/spotify";
 import { create } from "zustand";
 
-export type Track = {
-  id: string;
-  name: string;
-  artists: string[];
-  image: string;
-  durationMs: number;
-  uri: string;
-  previewUrl?: string;
-  uniqueKey?: string;
-};
-
-export type RepeatMode = "off" | "context" | "track";
-
-type PlayerState = {
-  playerInstance: Spotify.Player | null;
-  currentTrack: Track | null;
-  queue: Track[];
-  currentIndex: number;
-
-  isReady: boolean; // Spotify SDK ready
-  deviceId: string | null;
-  isPlaying: boolean;
-
-  isLoadingTrack: boolean; // 곡 정보 불러오기 중
-  setIsLoadingTrack: (loading: boolean) => void;
-
-  volume: number;
-  isShuffled: boolean;
-  repeatMode: RepeatMode;
-
-  position: number;
-  duration: number;
-  isQueueOpen: boolean;
-
-  setPlayerInstance: (player: Spotify.Player | null) => void;
-
-  optimisticPlay: (tracks: Track[], index: number) => void;
-  togglePlay: () => Promise<void>;
-  nextTrack: () => Promise<void>;
-  prevTrack: () => Promise<void>;
-  seekTo: (pos: number) => Promise<void>;
-
-  setVolume: (val: number) => Promise<void>;
-  toggleShuffle: (token: string) => Promise<void>;
-  cycleRepeatMode: (token: string) => Promise<void>;
-
-  syncStateFromSdk: (state: Spotify.PlaybackState) => void;
-
-  setQueue: (tracks: Track[]) => void;
-  toggleQueue: () => void;
-  setDeviceId: (id: string | null) => void;
-  setIsPlaying: (state: boolean) => void;
-  setIsReady: (ready: boolean) => void;
-  setPosition: (pos: number) => void;
-  setDuration: (dur: number) => void;
-};
-
 export const usePlayerStore = create<PlayerState>((set, get) => ({
+  // --- [상태 초기값] ---
   playerInstance: null,
   currentTrack: null,
   queue: [],
@@ -68,23 +13,39 @@ export const usePlayerStore = create<PlayerState>((set, get) => ({
   isReady: false,
   isPlaying: false,
   isQueueOpen: false,
-
   volume: 0.5,
   isShuffled: false,
   repeatMode: "off",
-
   duration: 0,
   position: 0,
-
   isLoadingTrack: false,
-  setIsLoadingTrack: (loading) => set({ isLoadingTrack: loading }),
 
+  // --- [상태 변경 단순 액션들] ---
+  setIsLoadingTrack: (loading) => set({ isLoadingTrack: loading }),
   setPlayerInstance: (player) => set({ playerInstance: player }),
+  setQueue: (tracks) => {
+    set({
+      queue: tracks,
+      currentIndex: 0,
+    });
+  },
+  toggleQueue: () =>
+    set((state) => ({
+      isQueueOpen: !state.isQueueOpen,
+    })),
+  setDeviceId: (id) => set({ deviceId: id }),
+  setIsReady: (ready) => set({ isReady: ready }),
+  setIsPlaying: (isPlaying) => set({ isPlaying }),
+  setPosition: (pos) => set({ position: pos }),
+  setDuration: (dur) => set({ duration: dur }),
+
+  // --- [비즈니스 로직 및 API 통신] ---
 
   togglePlay: async () => {
     const { isPlaying, playerInstance } = get();
     if (!playerInstance) return;
 
+    // 낙관적 업데이트: UI 먼저 즉각 변경 후 SDK 명령
     set({ isPlaying: !isPlaying });
     await playerInstance.togglePlay();
   },
@@ -93,27 +54,34 @@ export const usePlayerStore = create<PlayerState>((set, get) => ({
     const { playerInstance, queue, currentIndex, isShuffled } = get();
     if (!playerInstance || queue.length === 0) return;
 
-    set({ isLoadingTrack: true });
+    // [FLOW 0] 다음 곡 클릭시 로딩 UI(스피너/닷)
+    // set({ isLoadingTrack: true }); ux를 위해 셔플모드에서만
 
-    if (!isShuffled) {
-      const nextIndex = currentIndex + 1;
-      const nextTrack = queue[nextIndex] ?? null;
+    try {
+      if (!isShuffled) {
+        // [FLOW 1-A] 일반 모드: 다음 곡이 뭔지 아니까 UI 정보 즉시 변경
+        const nextIndex = currentIndex + 1;
+        if (nextIndex < queue.length) {
+          set({
+            currentIndex: nextIndex,
+            currentTrack: queue[nextIndex] ?? null,
+            position: 0,
+            duration: 0,
+            isPlaying: true,
+          });
+        }
+      } else {
+        // [FLOW 1-B] 셔플 모드: 스포티파이 서버가 무슨 곡을 줄지 모름
+        // 앨범 아트는 놔두고 프로그레스 바(position)만 0으로 밀어서 곡이 전환 중임을 암시
+        set({ isLoadingTrack: true, position: 0 });
+      }
 
-      if (nextIndex < queue.length) {
-        set({
-          currentIndex: nextIndex,
-          currentTrack: nextTrack,
-          position: 0,
-          duration: 0,
-          isPlaying: true,
-        });
-      } // queue의 끝이면?
-    } else {
-      // 셔플 모드일 때는 "어떤 곡이 나올지 모르니" 곡 정보는 건드리지 않습니다.
-      // 대신 사용자가 '다음'을 눌렀다는 피드백을 주기 위해 프로그레스 바만 0으로 밀어버립니다.
-      set({ position: 0 });
+      // [FLOW 2] 스포티파이 서버로 다음 곡 재생 명령 전달 (로딩은 syncStateFromSdk에서 꺼짐)
+      await playerInstance.nextTrack();
+    } catch (e) {
+      console.error("다음 곡 넘기기 실패:", e);
+      set({ isLoadingTrack: false });
     }
-    await playerInstance.nextTrack();
   },
 
   prevTrack: async () => {
@@ -126,27 +94,28 @@ export const usePlayerStore = create<PlayerState>((set, get) => ({
       return;
     }
 
-    set({ isLoadingTrack: true });
+    // set({ isLoadingTrack: true });
 
-    if (!isShuffled) {
-      const prevIndex = currentIndex - 1;
-      const prevTrack = queue[prevIndex] ?? null;
-
-      if (prevIndex >= 0) {
-        set({
-          currentIndex: prevIndex,
-          currentTrack: prevTrack,
-          position: 0,
-          duration: 0,
-          isPlaying: true,
-        });
+    try {
+      if (!isShuffled) {
+        const prevIndex = currentIndex - 1;
+        if (prevIndex >= 0) {
+          set({
+            currentIndex: prevIndex,
+            currentTrack: queue[prevIndex] ?? null,
+            position: 0,
+            duration: 0,
+            isPlaying: true,
+          });
+        }
+      } else {
+        set({ isLoadingTrack: true, position: 0 });
       }
-    } else {
-      // 셔플 모드: 곡 정보는 바꾸지 않고 로딩 상태(position=0)만 트리거
-      set({ position: 0 });
+      await playerInstance.previousTrack();
+    } catch (e) {
+      console.error("이전 곡 넘기기 실패:", e);
+      set({ isLoadingTrack: false });
     }
-
-    await playerInstance.previousTrack();
   },
 
   seekTo: async (pos) => {
@@ -158,15 +127,15 @@ export const usePlayerStore = create<PlayerState>((set, get) => ({
   },
 
   optimisticPlay: (tracks, index) => {
-    const trackToPlay = tracks[index] ?? null;
     set({
       queue: tracks,
       currentIndex: index,
-      currentTrack: trackToPlay,
+      currentTrack: tracks[index] ?? null,
       isPlaying: true,
     });
   },
 
+  // [SDK 이벤트 리스너] 스포티파이 서버에서 상태 변경 알림이 올 때마다 실행됨 (타 기기에서 변경 대비)
   syncStateFromSdk: (state) => {
     const { currentTrack, queue } = get();
     const sdkTrack = state.track_window.current_track;
@@ -187,6 +156,7 @@ export const usePlayerStore = create<PlayerState>((set, get) => ({
     set({
       isShuffled: state.shuffle,
       repeatMode: repeatModes[state.repeat_mode] ?? "off",
+      // [FLOW 4] 서버에서 새 곡 정보를 정상적으로 받아왔으므로 로딩 UI를 끔
       isLoadingTrack: false,
     });
   },
@@ -194,13 +164,17 @@ export const usePlayerStore = create<PlayerState>((set, get) => ({
   setVolume: async (val) => {
     const { playerInstance } = get();
     set({ volume: val });
+
     if (playerInstance) {
+      // ux를 위해 playerInstance 없어도 바로 return X
       await playerInstance.setVolume(val);
     }
   },
 
   toggleShuffle: async (token) => {
     const { isShuffled, deviceId } = get();
+    // 셔플, 반복 기능은 sdk가 아닌 web API에서 다루기 때문에
+    // playerInstance 아닌 deviceId 필요
     if (!deviceId) return;
 
     const nextState = !isShuffled;
@@ -209,7 +183,7 @@ export const usePlayerStore = create<PlayerState>((set, get) => ({
     try {
       await setShuffle(nextState, deviceId, token);
     } catch (e) {
-      console.error("Shuffle Error:", e);
+      console.error("셔플 모드 변경 실패:", e);
       set({ isShuffled: !nextState });
     }
   },
@@ -227,24 +201,8 @@ export const usePlayerStore = create<PlayerState>((set, get) => ({
     try {
       await setRepeatMode(nextMode, deviceId, token);
     } catch (e) {
-      console.error("Repeat Error:", e);
+      console.error("반복 모드 변경 실패:", e);
       set({ repeatMode });
     }
   },
-
-  setQueue: (tracks) => {
-    set({
-      queue: tracks,
-      currentIndex: 0,
-    });
-  },
-  toggleQueue: () =>
-    set((state) => ({
-      isQueueOpen: !state.isQueueOpen,
-    })),
-  setDeviceId: (id) => set({ deviceId: id }),
-  setIsReady: (ready) => set({ isReady: ready }),
-  setIsPlaying: (isPlaying) => set({ isPlaying }),
-  setPosition: (pos) => set({ position: pos }),
-  setDuration: (dur) => set({ duration: dur }),
 }));
